Move the color config subscription into FactionAPI
