+++ projects/faction/src/main/java/de/teamlapen/faction/common/config/FactionConfig.java
@@ -3,6 +3,7 @@ package de.teamlapen.faction.common.config;
 import de.teamlapen.faction.FactionsMod;
 import de.teamlapen.faction.Services;
 import de.teamlapen.faction.client.config.ClientConfig;
+import de.teamlapen.faction.client.config.values.ColorConfigValue;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.config.IConfigSpec;
@@ -88,7 +89,7 @@ public class FactionConfig extends Services {
     public record Config<T extends IConfigs>(T config, ModConfigSpec spec) {
 
         public static <T extends IConfigs> Config<T> create(Function<ModConfigSpec.Builder, T> consumer) {
-            var builder = new ModConfigSpec.Builder().configure(consumer);
+            var builder = ColorConfigValue.configure(consumer);
             return new Config<>(builder.getLeft(), builder.getRight());
         }
