+++ projects/vampirism/src/main/java/de/teamlapen/vampirism/common/config/ModConfig.java
@@ -2,6 +2,7 @@ package de.teamlapen.vampirism.common.config;
 
 
 import de.teamlapen.faction.Services;
+import de.teamlapen.faction.client.config.values.ColorConfigValue;
 import de.teamlapen.vampirism.VampirismMod;
 import de.teamlapen.vampirism.api.ThreadSafeAPI;
 import de.teamlapen.vampirism.client.config.ClientConfig;
@@ -123,7 +124,6 @@ public class ModConfig extends Services {
         if (configEvent.getConfig().getSpec() == balanceSpec) {
             helper.onBalanceConfigChanged(configEvent);
         }
-        ColorConfigValue.reloadAll(configEvent.getConfig().getSpec());
     }
 
     public void onReload(final ModConfigEvent.Reloading configEvent) {
@@ -133,7 +133,6 @@ public class ModConfig extends Services {
         if (configEvent.getConfig().getSpec() == balanceSpec) {
             helper.onBalanceConfigChanged(configEvent);
         }
-        ColorConfigValue.reloadAll(configEvent.getConfig().getSpec());
     }
 
     //</editor-fold>
@@ -157,11 +156,8 @@ public class ModConfig extends Services {
     private record Config<T>(T config, ModConfigSpec spec) {
 
         public static <T> Config<T> create(Function<ModConfigSpec.Builder, T> consumer) {
-            ColorConfigValue.beginCollection();
-            var builder = new ModConfigSpec.Builder().configure(consumer);
-            ModConfigSpec spec = builder.getRight();
-            ColorConfigValue.endCollection(spec);
-            return new Config<>(builder.getLeft(), spec);
+            var builder = ColorConfigValue.configure(consumer);
+            return new Config<>(builder.getLeft(), builder.getRight());
         }
 
         public boolean isSpec(IConfigSpec spec) {
